wip
- added `htmelt` global client namespace to watch mode, which exposes functions for defining module exports at runtime, so they can be imported and updated by modules generated with the `config.loadDevModule` function

- disable automatic page reloads. this will be brought back once I have the time to add module graph tracking to watch mode, which allows us to determine hmr boundaries

- added `config.fsAllowedDirs` and populate it automatically with the realpaths of linked dependencies

- fixed sourcemap `sources` array for out-of-root modules

- the `config.esbuild.plugins` array will always be defined now

- started using esbuild’s incremental build API for <script> bundling

- disabled esbuild tree-shaking in watch mode

- rebuilds now happen even if changed file is accepted by hmr instance, so that manual page refreshes are sent the latest revisions

- added support for /@fs/ prefix to dev server, so out-of-root modules can have their sourcemaps loaded

- rebuild when a module from a linked dependency is changed

- dev server requests are delayed until the most recent rebuild is completed

- watched files must exist in `config.modules` in order to trigger a rebuild when changed

- disable rebuilds in response to "add" watcher events. this will be brought back when I have time to add tracking of failed module resolutions

- the `config.watchFiles` array will always be defined now

- enable the websocket server even if no hmr plugins exist, so that rebuilds can trigger page refreshes

+++ src/cli.mts
@@ -12,6 +12,8 @@ import {
   ServePlugin,
 } from '@htmelt/plugin'
 import cac from 'cac'
+import { FSWatcher } from 'chokidar'
+import * as esbuild from 'esbuild'
 import * as fs from 'fs'
 import { cyan, red, yellow } from 'kleur/colors'
 import mitt, { Emitter } from 'mitt'
@@ -33,11 +35,13 @@ import { buildHTML, parseHTML } from './html.mjs'
 import {
   baseRelative,
   createDir,
+  isEqualUnordered,
   loadBundleConfig,
   lowercaseKeys,
+  resolveDevMapSources,
 } from './utils.mjs'
 
-const cli = cac('html-bundle')
+const cli = cac('htmelt')
 
 cli
   .command('')
@@ -64,19 +68,26 @@ async function bundle(config: Config, flags: Flags) {
     server = await installHttpServer(config, servePlugins)
   }
 
-  type HTMLEntry = {
+  type EntryMetadata = {
     document: ParentNode
     scripts: ScriptReference[]
+    bundle: ScriptBundle
   }
 
   type ScriptBundle = {
-    hmr?: boolean
-    entries: Set<string>
+    id: string
+    hmr: boolean
+    scripts: Set<string>
+    importers: Set<Entry>
+    context?: esbuild.BuildContext
+    metafile?: esbuild.Metafile
   }
 
+  const scriptBundles = new Map<string, ScriptBundle>()
+
   const build = async () => {
-    const htmlEntries = new Map<Entry, HTMLEntry>()
-    const scriptBundles = new Map<string, ScriptBundle>()
+    const htmlEntries = new Map<Entry, EntryMetadata>()
+    const scriptEntriesByBundle = new Map<ScriptBundle, string[]>()
 
     const seen = new Set<string>()
     for (const entry of config.entries) {
@@ -86,45 +97,54 @@ async function bundle(config: Config, flags: Flags) {
       if (seen.has(key)) continue
       seen.add(key)
 
-      const scriptBundle = scriptBundles.get(bundleId) || { entries: new Set() }
-      scriptBundles.set(bundleId, scriptBundle)
+      let bundle = scriptBundles.get(bundleId)
+      if (!bundle) {
+        bundle = {
+          id: bundleId,
+          hmr: true,
+          scripts: new Set(),
+          importers: new Set(),
+        }
+        scriptBundles.set(bundleId, bundle)
+      } else if (!scriptEntriesByBundle.has(bundle)) {
+        scriptEntriesByBundle.set(bundle, [...bundle.scripts])
+        bundle.scripts.clear()
+        bundle.hmr = true
+      }
+
       if (entry.hmr == false) {
-        scriptBundle.hmr = false
+        bundle.hmr = false
       }
 
       if (file.endsWith('.html')) {
         const html = fs.readFileSync(file, 'utf8')
         const document = parseHTML(html)
         const scripts = findRelativeScripts(document, file, config)
-        htmlEntries.set(entry, { document, scripts })
+        htmlEntries.set(entry, { document, scripts, bundle })
+        bundle.importers.add(entry)
         for (const script of scripts) {
-          scriptBundle.entries.add(script.srcPath)
+          bundle.scripts.add(script.srcPath)
         }
       } else if (/\.[mc]?[tj]sx?$/.test(file)) {
-        scriptBundle.entries.add(path.resolve(file))
+        bundle.scripts.add(path.resolve(file))
       } else {
         console.warn(red('⚠'), 'unsupported entry type:', file)
       }
     }
 
     await Promise.all([
-      ...Array.from(htmlEntries, ([entry, { document, scripts }]) =>
-        buildHTML(entry, document, scripts, config, flags)
-      ),
       Promise.all(
-        Array.from(scriptBundles, async ([bundleId, { hmr, entries }]) => {
-          const { metafile } = await buildEntryScripts(
-            [...entries],
-            config,
-            flags
-          )
-          const bundle: Plugin.Bundle = {
-            id: bundleId,
-            hmr,
-            entries,
-            ...metafile,
+        Array.from(scriptBundles, async ([bundleId, bundle]) => {
+          const oldEntries = scriptEntriesByBundle.get(bundle)
+          const newEntries = [...bundle.scripts]
+          if (!bundle.context || isEqualUnordered(oldEntries!, newEntries)) {
+            bundle.context = await buildEntryScripts(newEntries, config, flags)
           }
-          return [bundleId, bundle] as const
+
+          const { metafile } = await bundle.context!.rebuild()
+          bundle.metafile = metafile!
+
+          return [bundleId, bundle as Plugin.Bundle] as const
         })
       )
         .then(Object.fromEntries<Plugin.Bundle>)
@@ -133,6 +153,9 @@ async function bundle(config: Config, flags: Flags) {
             plugin.bundles?.(bundles)
           }
         }),
+      ...Array.from(htmlEntries, ([entry, { document, scripts, bundle }]) =>
+        buildHTML(entry, document, scripts, config, flags)
+      ),
       ...config.scripts.map(async entry => {
         console.log(yellow('⌁'), baseRelative(entry))
         return compileSeparateEntry(entry, config).then(async code => {
@@ -149,7 +172,7 @@ async function bundle(config: Config, flags: Flags) {
   }
 
   const timer = performance.now()
-  await build()
+  await (config.lastBuild = build())
   console.log(
     cyan('build complete in %sms'),
     (performance.now() - timer).toFixed(2)
@@ -168,28 +191,42 @@ async function bundle(config: Config, flags: Flags) {
     const hmrInstances: Plugin.HmrInstance[] = []
 
     const hmrPlugins = config.plugins.filter(p => p.hmr) as HmrPlugin[]
-    if (hmrPlugins.length) {
-      await installWebSocketServer(server, config, hmrPlugins, hmrInstances)
-    }
+    const clients = installWebSocketServer(
+      server,
+      config,
+      hmrPlugins,
+      hmrInstances
+    )
 
     const watcher = config.watcher!
     const changedFiles = new Set<string>()
 
-    config.watchFiles?.forEach(file => {
+    registerLinkedPackages(watcher, config.fsAllowedDirs)
+
+    config.watchFiles.forEach(file => {
       watcher.add(path.resolve(file))
     })
 
-    watcher.on('add', async file => {
-      await rebuild()
+    // TODO: track failed module resolutions and only rebuild if a file
+    // is added that matches one of them.
+    /*watcher.on('add', async file => {
+      await scheduleRebuild()
       console.log(cyan('+'), file)
-    })
+    })*/
 
     watcher.on('change', async file => {
-      changedFiles.add(file)
-      await rebuild()
+      file = baseRelative(path.resolve(file))
+      if (file in config.modules!) {
+        changedFiles.add(file)
+        await requestRebuild()
+      }
     })
 
     watcher.on('unlink', async file => {
+      // Absolute files are typically not added to the build directory.
+      if (path.isAbsolute(file)) {
+        return
+      }
       const outPath = config.getBuildPath(file).replace(/\.[jt]sx?$/, '.js')
       try {
         fs.rmSync(outPath)
@@ -200,18 +237,34 @@ async function bundle(config: Config, flags: Flags) {
           fs.rmSync(outDir)
           outDir = path.dirname(outDir)
         }
+        console.log(red('–'), file)
       } catch {}
-      console.log(red('–'), file)
     })
 
-    const rebuild = debounce(async () => {
+    const requestRebuild = debounce(() => {
+      config.lastBuild = rebuild()
+    }, 200)
+
+    const rebuild = async () => {
       console.clear()
 
-      let needRebuild = false
+      let isFullReload = false
+      const fullReloadFiles = new Set<string>()
+      for (const bundle of scriptBundles.values()) {
+        if (!bundle.hmr) {
+          for (const srcPath in bundle.metafile!.inputs) {
+            fullReloadFiles.add(srcPath)
+          }
+        }
+      }
 
       const acceptedFiles = new Map<Plugin.HmrInstance, string[]>()
-      accept: for (const file of changedFiles) {
+      accept: for (let file of changedFiles) {
         console.log(cyan('↺'), file)
+        if (fullReloadFiles.has(file)) {
+          isFullReload = true
+          break
+        }
         for (const hmr of hmrInstances) {
           if (hmr.accept(file)) {
             let files = acceptedFiles.get(hmr)
@@ -222,39 +275,70 @@ async function bundle(config: Config, flags: Flags) {
             continue accept
           }
         }
-        needRebuild = true
+        isFullReload = true
         break
       }
+
       changedFiles.clear()
+      await Promise.all(
+        Array.from(acceptedFiles, ([hmr, files]) => hmr.update(files))
+      )
 
-      if (needRebuild) {
-        try {
-          config.events.emit('will-rebuild')
-          const timer = performance.now()
-          await build()
-          config.events.emit('rebuild')
+      try {
+        config.events.emit('will-rebuild', { isFullReload })
+        const timer = performance.now()
+        await build()
+        config.events.emit('rebuild', { isFullReload })
 
-          for (const plugin of config.plugins) {
-            if (!plugin.buildEnd) continue
-            await plugin.buildEnd(true)
-          }
-
-          console.log(
-            cyan('build complete in %sms'),
-            (performance.now() - timer).toFixed(2)
-          )
-        } catch (e: any) {
-          console.error(e)
+        for (const plugin of config.plugins) {
+          if (!plugin.buildEnd) continue
+          await plugin.buildEnd(true)
         }
-      } else {
-        await Promise.all(
-          Array.from(acceptedFiles, ([hmr, files]) => hmr.update(files))
+
+        console.log(
+          cyan('build complete in %sms'),
+          (performance.now() - timer).toFixed(2)
         )
+
+        if (isFullReload) {
+          await Promise.all(Array.from(clients, client => client.reload()))
+        }
+      } catch (e: any) {
+        console.error(e)
       }
+
       console.log(yellow('watching files...'))
-    }, 200)
+    }
 
     console.log(yellow('watching files...'))
+  }
+}
+
+// This function adds all linked packages to the watcher
+// so that the watcher will detect changes in these packages.
+function registerLinkedPackages(watcher: FSWatcher, fsAllowedDirs: string[]) {
+  const nodeModulesDir = path.resolve('node_modules')
+  const nodeModules = fs
+    .readdirSync(nodeModulesDir)
+    .flatMap(name =>
+      name[0] === '@'
+        ? fs
+            .readdirSync(path.join(nodeModulesDir, name))
+            .map(scopedName => path.join(name, scopedName))
+        : name
+    )
+
+  for (const name of nodeModules) {
+    const dependencyDir = path.join(nodeModulesDir, name)
+    const resolvedDependencyDir = fs.realpathSync(dependencyDir)
+    if (
+      resolvedDependencyDir !== dependencyDir &&
+      path.relative(process.cwd(), resolvedDependencyDir).startsWith('..')
+    ) {
+      console.log('linked dep:', name, '->', resolvedDependencyDir)
+      fsAllowedDirs.push(resolvedDependencyDir)
+      watcher.add(resolvedDependencyDir)
+    }
   }
 }
 
@@ -265,7 +349,7 @@ async function installHttpServer(config: Config, servePlugins: ServePlugin[]) {
     createServer = (await import('https')).createServer
     serverOptions = config.server.https
     if (!serverOptions.cert) {
-      const cert = await getCertificate('node_modules/.html-bundle/self-signed')
+      const cert = await getCertificate('node_modules/.htmelt/self-signed')
       serverOptions.cert = cert
       serverOptions.key = cert
     }
@@ -275,11 +359,15 @@ async function installHttpServer(config: Config, servePlugins: ServePlugin[]) {
   }
 
   // The dev server allows access to files within these directories.
-  const fsAllow = new RegExp(`^/(${[config.build, config.assets].join('|')})/`)
+  const fsAllowRE = new RegExp(
+    `^/(${[config.build, config.assets].join('|')})/`
+  )
 
   const server = createServer(serverOptions, async (req, response) => {
     const request = Object.assign(req, parseURL(req.url!)) as Plugin.Request
     request.searchParams = new URLSearchParams(request.search || '')
+
+    await config.lastBuild
 
     let file: Plugin.VirtualFileData | null = null
     for (const plugin of servePlugins) {
@@ -290,7 +378,8 @@ async function installHttpServer(config: Config, servePlugins: ServePlugin[]) {
 
     // If no plugin handled the request, check the virtual filesystem.
     if (!file) {
-      const uri = request.pathname
+      let uri = request.pathname
+      let filePath: string
 
       let virtualFile = config.virtualFiles[uri]
       if (virtualFile) {
@@ -300,14 +389,36 @@ async function installHttpServer(config: Config, servePlugins: ServePlugin[]) {
         file = await virtualFile
       }
 
+      const isFileRequest = uri.startsWith('/@fs/')
+      if (isFileRequest) {
+        filePath = uri.slice(4)
+      } else {
+        filePath = path.join(process.cwd(), uri)
+      }
+
       // If no virtual file exists, check the local filesystem.
-      if (!file && fsAllow.test(uri)) {
-        try {
-          console.log(cyan('reading'), '.' + uri)
-          file = {
-            data: fs.readFileSync('.' + uri),
-          }
-        } catch {}
+      if (!file) {
+        let isAllowed = false
+        if (isFileRequest) {
+          isAllowed = config.fsAllowedDirs.some(
+            dir => !path.relative(dir, filePath).startsWith('..')
+          )
+        } else {
+          isAllowed = fsAllowRE.test(uri)
+        }
+        if (isAllowed) {
+          try {
+            file = {
+              data: fs.readFileSync(filePath),
+            }
+          } catch {}
+        }
+      }
+
+      if (file && uri.endsWith('.map')) {
+        const map = JSON.parse(file.data.toString('utf8'))
+        resolveDevMapSources(map, process.cwd(), path.dirname(filePath))
+        file.data = JSON.stringify(map)
       }
     }
 
@@ -342,7 +453,7 @@ async function installHttpServer(config: Config, servePlugins: ServePlugin[]) {
   return server
 }
 
-async function installWebSocketServer(
+function installWebSocketServer(
   server: import('http').Server,
   config: Config,
   hmrPlugins: HmrPlugin[],
@@ -395,15 +506,19 @@ async function installWebSocketServer(
       }
       return evaluate(this, path)
     }
-    async evaluateModule(file: string, args?: any[]) {
-      const moduleUrl = new URL(file, import.meta.url)
+    async evaluateModule(file: string | URL, args?: any[]) {
+      const moduleUrl =
+        typeof file === 'string' ? new URL(file, import.meta.url) : file
       const mtime = fs.statSync(moduleUrl).mtimeMs
 
       const path = `/${md5Hex(moduleUrl.href)}.${mtime}.js`
       if (config.virtualFiles[path] == null) {
         let compiled = compiledModules.get(moduleUrl.href)
         if (compiled?.mtime != mtime) {
-          const data = await compileSeparateEntry(file, config, 'esm')
+          const entry = decodeURIComponent(moduleUrl.pathname)
+          const data = await compileSeparateEntry(entry, config, {
+            format: 'esm',
+          })
           compiledModules.set(
             moduleUrl.href,
             (compiled = {
@@ -470,6 +585,8 @@ async function installWebSocketServer(
       client,
     })
   })
+
+  return clients
 }
 
 async function getCertificate(cacheDir: string) {
